feat(FN-1855): move SqlDbDataSource to separate sub-directory. Add more entities
