+++ libs/common/package.json
@@ -17,13 +17,16 @@
   },
   "license": "MIT",
   "author": "UK Export Finance",
+  "exports": {
+    ".": "./src/index.ts",
+    "./sql-db-connection": "./src/sql-db-connection/index.ts"
+  },
   "main": "src/index.ts",
-  "types": "./src/index.ts",
   "scripts": {
-    "db:drop-all": "npm run typeorm -- schema:drop -d src/sql-db/data-source.ts",
-    "db:generate-migration": "cross-var npm run typeorm -- migration:generate -d src/sql-db/data-source.ts -p src/sql-db/migrations/$npm_config_name",
-    "db:migrate": "npm run typeorm -- migration:run -d src/sql-db/data-source.ts",
-    "db:migrate:down": "npm run typeorm -- migration:revert -d src/sql-db/data-source.ts",
+    "db:drop-all": "npm run typeorm -- schema:drop -d src/sql-db-connection/data-source.ts",
+    "db:generate-migration": "cross-var npm run typeorm -- migration:generate -d src/sql-db-connection/data-source.ts -p src/sql-db-connection/migrations/$npm_config_name",
+    "db:migrate": "npm run typeorm -- migration:run -d src/sql-db-connection/data-source.ts",
+    "db:migrate:down": "npm run typeorm -- migration:revert -d src/sql-db-connection/data-source.ts",
     "db:reset": "npm run db:drop-all && npm run db:migrate",
     "lint": "eslint \"src/**/*.{ts,js}\" && npm run type-check",
     "lint:fix": "eslint \"src/**/*.{ts,js}\"",
