+++ libs/common/src/constants/currency.ts
@@ -1,0 +1,6 @@
+export const CURRENCY = {
+  GBP: 'GBP',
+  EUR: 'EUR',
+  USD: 'USD',
+  JPY: 'JPY',
+} as const;
