+++ libs/common/src/constants/index.ts
@@ -1,1 +1,2 @@
+export * from './currency';
 export * from './utilisation-reports';
