+++ libs/common/src/index.ts
@@ -1,4 +1,5 @@
 export * from './constants';
 export * from './helpers';
-export * from './sql-db';
+// './sq-db-connection' should not be exported from here - see sql-db.md for details
+export * from './sql-db-entity';
 export * from './types';
