+++ libs/common/src/sql-db-connection/Dockerfile
