+++ libs/common/src/sql-db-connection/index.ts
@@ -1,0 +1,1 @@
+export { SqlDbDataSource } from './data-source';
