+++ libs/common/src/sql-db-connection/init.sh
