+++ libs/common/src/sql-db-entity/partials/index.ts
