+++ libs/common/src/sql-db-entity/partials/report-period.ts
