--- libs/common/src/sql-db/entity/index.ts
@@ -1,1 +1,0 @@
-export * from './utilisation-report';
