--- libs/common/src/sql-db/index.ts
@@ -1,2 +1,0 @@
-export * from './entity';
-export * from './data-source';
