+++ libs/common/src/types/currency.ts
@@ -1,0 +1,4 @@
+import { ValuesOf } from './types-helper';
+import { CURRENCY } from '../constants';
+
+export type Currency = ValuesOf<typeof CURRENCY>;
