+++ libs/common/src/types/index.ts
@@ -1,2 +1,3 @@
+export * from './currency';
 export * from './types-helper';
 export * from './utilisation-reports';
